% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.fit.R
\name{apt.fit}
\alias{apt.fit}
\title{To fit the negative of the first derivative data on the basis of the Gaussian fit
(two Gaussian Fit)}
\usage{
apt.fit(x, m1 = 44, sd1 = 1, A1 = 3, m2 = 53, sd2 = 1, A2 = 3,
  method = "BFGS", cvalue = 2e-09, temp, plottitle, ylabel, xlabel)
}
\arguments{
\item{x}{the data to be fitted. These data are the first derivative data. See the examples
to extract the first derivative data to be passed for the two Gaussian fitting.}

\item{m1}{Initial value of mu1 (first mu). This is the initial value of the first Gaussian mean
value of temperature. The default value is 44.}

\item{sd1}{Initial value of first standard deviation. The deault value is 1.}

\item{A1}{Initial value of first amplitude. The default value is 3.}

\item{m2}{Initial value of mu2 (second mu). This is the initial value of the second Gaussian mean
value of temperature. The default value is 53.}

\item{sd2}{Initial value of second standard deviation. The default value is 1.}

\item{A2}{Initial value of  second amplitude. The default value is A2.}

\item{method}{The method to be implemented. The default method is "BFGS". For more details
see also \code{\link[stats]{optim}}.}

\item{cvalue}{This is the tolerance value (reltol value).The default value is 2e-9.
For more details see also \code{\link[stats]{optim}}.}

\item{temp}{These are the temperature values. See the examples to extract the temperature values
to be passed for the two Gaussian fitting of the data.}

\item{plottitle}{The tittle of the plot to be defined by the user, only in the case if the
user wants to generate the plot. If missing the plotttile in the arguments, then, the results
will be generated in the form of the Class \code{data.frame}.}

\item{ylabel}{ylabel should be defined only if the user defined the plottitle. The default is
-dF/dT.}

\item{xlabel}{The xlabel should be defined only if the user defined the plottitle. The default
is "Temperature (°C)".}
}
\value{
Return an object of Class \code{data.frame}. Also, return the results in the form of a
plot.
}
\description{
Function \code{apt.fit} acts on the object of Class \code{data.frame} and implements
the function \code{\link[stats]{optim}} for fitting the data on the basis of two Gaussian fit. The
function \code{apt.fit} saves the fitting results in an object of Class \code{data.frame}. The
function \code{apt.fit} also produces the fitting plots, which are generated on the basis of the
function \code{\link[ggplot2]{ggplot}}.
}
\details{
Function \code{apt.fit} acts on the object of Class \code{data.frame} and implements
the function \code{\link[stats]{optim}} for fitting the data on the basis of two Gaussian fit.
The function \code{apt.fit} saves the fitting results in an object of Class \code{data.frame}.
The function \code{apt.fit} also produces the fitting plots, which are generated on the basis of
the function \code{\link[ggplot2]{ggplot}}. To have an idea about the initial value of the
Gaussian parameters m1,A1,m2 and A2 regarding fitting, we recommend to have a look on the results of the
plot of negative of first derivative, generated through the function \code{\link[AptaqPCR]{apt.plot}}.
For more details see the vignettes of the package \code{AptaqPCR}. See also \code{\link[AptaqPCR]{ReadAptaqPCR}},
\code{\link[AptaqPCR]{mean.apt}}, \code{\link[AptaqPCR]{diff.apt}}, \code{\link[AptaqPCR]{apt.plot}} and
\code{\link[AptaqPCR]{apt.fit.value}}.
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "multiapt31.csv", package = "AptaqPCR")
read.datacsv <- ReadAptaqPCR(file1)

## To combine the replicates on the basis of the mean
meandata <- mean.apt(read.datacsv)

## To compute the first derivative of the combined replicates
diffdata <- diff.apt(meandata)

## To express the first derivative data
data <- slot(diffdata,"diffData")
temp <- data[,"temp"]

## To fit the negative of first derivative data to generate the fitting results in the
##form of values (an object of Class data.frame)

nc <- apt.fit(data[,"nc"],44,1,3,53,1,3,"BFGS",2e-9,temp)
b <- apt.fit(data[,"b"],44,1,3,53,1,3,"BFGS",2e-9,temp)
c <- apt.fit(data[,"c"],44,1,3,53,1,3,"BFGS",2e-9,temp)
d <- apt.fit(data[,"d"],44,1,3,53,1,3,"BFGS",2e-9,temp)
e <- apt.fit(data[,"e"],44,1,3,53,1,3,"BFGS",2e-9,temp)
f <- apt.fit(data[,"f"],44,1,3,53,1,3,"BFGS",2e-9,temp)
g <- apt.fit(data[,"g"],44,1,3,53,1,3,"BFGS",2e-9,temp)
h <- apt.fit(data[,"h"],44,1,3,53,1,3,"BFGS",2e-9,temp)


## To fit the negative of the first derivative data to generate the fitting results in the
## form of plots, where the names of plos are defined on the basis of the concentration

nc.plot <- apt.fit(data[,"nc"],44,1,3,53,1,3,"BFGS",2e-9,temp,"0 pM")
b.plot <- apt.fit(data[,"b"],44,1,3,53,1,3,"BFGS",2e-9,temp,"30.50 pM")
c.plot <- apt.fit(data[,"c"],44,1,3,53,1,3,"BFGS",2e-9,temp,"15.25 pM")
d.plot <- apt.fit(data[,"d"],44,1,3,53,1,3,"BFGS",2e-9,temp,"7.62 pM")
e.plot <- apt.fit(data[,"e"],44,1,3,53,1,3,"BFGS",2e-9,temp,"3.81 pM")
f.plot <- apt.fit(data[,"f"],44,1,3,53,1,3,"BFGS",2e-9,temp,"1.90 pM")
g.plot <- apt.fit(data[,"g"],44,1,3,53,1,3,"BFGS",2e-9,temp,"0.95 pM")
h.plot <- apt.fit(data[,"h"],44,1,3,53,1,3,"BFGS",2e-9,temp,"0.47 pM")

## To arrange all the plots in a single plot
grid.arrange(nc.plot, b.plot, c.plot, d.plot, e.plot, f.plot, g.plot, h.plot, ncol = 2, nrow = 4)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{ReadAptaqPCR,}
\keyword{apt.fit}
\keyword{aptaqpcr,}
\keyword{diff.apt,}
\keyword{mean.apt,}
