% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{aptaqpcr-class}
\alias{aptaqpcr-class}
\title{S4 Class \code{aptaqpcr} to contain the data}
\section{Slots}{

\describe{
\item{\code{initialData}}{contains the initial data in the form of an object.
of Class \code{data.frame}. These are the raw fluorescent data.}

\item{\code{meanData}}{contains the result of the function \code{mean.apt} in the form of an object
of Class \code{data.frame}. These are the data obtained after combining replicates.}

\item{\code{diffData}}{contains the result of the function \code{diff.apt} in the form of an
object of Class \code{data.frame}. These are the first derivative data of the combined
replicates.}

\item{\code{fitData}}{contains the result of the functions \code{apt.fit.value} and \code{apt.fit.all}
in the form of an object of Class \code{data.frame}. These are the fitting results obtained
from the two Gaussian fit.}

\item{\code{confMeanData}}{contains the result (the mean values) of the function \code{mean.conf}
in the form of an object of Class \code{data.frame}. These are Gaussian mean values of the
temperature.}

\item{\code{confData}}{contains the result (the confidence intervals) of the function \code{mean.conf}
in the form of an object of Class \code{data.frame}. These data are the confidence intervals
of the Gaussian mean values.}
}}

