% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AptaqPCR.gui.R
\name{AptaqPCR.gui}
\alias{AptaqPCR.gui}
\title{Run the graphical user interface for melting curve analysis for diagnostics}
\usage{
AptaqPCR.gui()
}
\description{
Run the graphical user interface
}
\examples{
AptaqPCR.gui()
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
