% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.apt.R
\docType{methods}
\name{diff.apt}
\alias{diff.apt}
\title{To compute the first derivative of the combined replicates}
\usage{
\S4method{diff.apt}{aptaqpcr}(aptaqpcr)
}
\arguments{
\item{aptaqpcr}{the object of Class \code{"aptaqpcr"}, produced as an output of the function
\code{mean.apt}.}
}
\value{
object of Class \code{"aptaqpcr"}.
}
\description{
Function \code{diff.apt} computes the first derivative of the combined
replicates and uses the data to populate an object of Class \code{"aptaqpcr"}.
}
\details{
Function \code{diff.apt} acts on an object of Class \code{"aptaqpcr"}, which are produced
as an output of the function \code{mean.apt} and uses the data to populate an object of
Class \code{"aptaqpcr"}. See also \code{\link[AptaqPCR]{ReadAptaqPCR}} and \code{\link[AptaqPCR]{mean.apt}}.
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "multiapt31.csv", package = "AptaqPCR")
read.datacsv <- ReadAptaqPCR(file1)

## To combine the replicates on the basis of the mean
meandata <- mean.apt(read.datacsv)

## To compute the first derivative of the combined replicates
diffdata <- diff.apt(meandata)

## To express all the data
diffdata

## To express the first derivative data
slot(diffdata,"diffData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{ReadAptaqPCR,}
\keyword{aptaqpcr,}
\keyword{diff.apt}
\keyword{mean.apt,}
