% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.apt.R
\docType{methods}
\name{mean.apt}
\alias{mean.apt}
\title{To combine the replicates of experimental data on the basis of mean}
\usage{
\S4method{mean.apt}{aptaqpcr}(aptaqpcr)
}
\arguments{
\item{aptaqpcr}{an object of Class \code{"aptaqpcr"}, produced as an output of the function
\code{ReadAptaqPCR}.}
}
\value{
object of Class \code{"aptaqpcr"}.
}
\description{
Function \code{mean.apt} combines the replicates and uses the data to populate
an object of Class \code{"aptaqpcr"}.
}
\details{
Function \code{mean.apt} acts on an object of Class \code{"aptaqpcr"}, produced
as an output of the function \code{ReadAptaqPCR}. It combines the replicates and uses
the data to populate an object of Class \code{"aptaqpcr"}.See also \code{\link[AptaqPCR]{ReadAptaqPCR}}.
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "multiapt31.csv", package = "AptaqPCR")
read.datacsv <- ReadAptaqPCR(file1)

##To combine the replicates on the basis of mean
meandata <- mean.apt(read.datacsv)

## To express all the data
meandata

## To express the combined replicates
slot(meandata,"meanData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{ReadAptaqPCR,}
\keyword{aptaqpcr,}
\keyword{mean.apt}
