% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAptaqPCR.R
\name{ReadAptaqPCR}
\alias{ReadAptaqPCR}
\title{To read the .csv as well as .txt file of experimental data}
\usage{
ReadAptaqPCR(file, type = ".csv")
}
\arguments{
\item{file}{the name of the file to read in.}

\item{type}{the type of the file out of ".csv" or ".txt" to read in. The default parameter
is ".csv".}
}
\value{
object of Class \code{"aptaqpcr"}.
}
\description{
Function \code{ReadAptaqPCR} reads in .csv as well as .txt files of raw fluorescent
melting temprature data of RTqPCR and uses the data to populate an object of Class \code{"aptaqpcr"}.
}
\details{
Allows the user to read in raw fluorescence data.
More details are provided in the \code{AptaqPCR} package vignette.
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "multiapt31.csv", package = "AptaqPCR")
read.datacsv <- ReadAptaqPCR(file1)

## To express all the data
read.datacsv

## To express the initial data (all the data read in, except concentration)
slot(read.datacsv,"initialData")

## To read the .txt file
file1 <- system.file("exData", "multiapt31.txt", package = "AptaqPCR")
read.datatxt <- ReadAptaqPCR(file1, type = ".txt")

## To express the initial data (except concentration)
slot(read.datatxt, "initialData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{ReadAptaqPCR}
\keyword{aptaqpcr,}
